+++ bahmnicore-omod/src/main/java/org/bahmni/module/bahmnicore/web/v1_0/controller/BahmniConfigController.java
@@ -6,9 +6,9 @@ import org.bahmni.module.admin.config.service.BahmniConfigService;
 import org.bahmni.module.bahmnicore.contract.drugorder.DrugOrderConfigResponse;
 import org.bahmni.module.bahmnicore.contract.encounter.data.ConceptData;
 import org.bahmni.module.bahmnicore.contract.encounter.response.EncounterConfigResponse;
-import org.bahmni.module.bahmnicore.contract.patient.response.PatientConfigResponse;
 import org.bahmni.module.bahmnicore.service.BahmniDrugOrderService;
-import org.bahmni.module.bahmnicore.service.BahmniPatientService;
+import org.bahmni.module.bahmnicommons.contract.patient.response.PatientConfigResponse;
+import org.bahmni.module.bahmnicommons.service.BahmniPatientService;
 import org.openmrs.Concept;
 import org.openmrs.EncounterType;
 import org.openmrs.OrderType;
