+++ pom.xml
@@ -48,7 +48,7 @@
         <idgenWebServicesVersion>1.3-SNAPSHOT</idgenWebServicesVersion>
         <providermanagementVersion>2.13.0</providermanagementVersion>
         <velocityVersion>1.6.2</velocityVersion>
-        <bahmniCommons.version>0.2-SNAPSHOT</bahmniCommons.version>
+        <bahmniCommons.version>1.0.0-SNAPSHOT</bahmniCommons.version>
 
         <!--test dependencies-->
         <junitVersion>4.13</junitVersion>
